--- src/Frontend/app/src/main/java/br/com/aula/text/Comentario.java
@@ -1,33 +1,0 @@
-package br.com.aula.text;
-
-public class Comentario {
-    private int id; // Atributo para armazenar o ID
-    private String nome;
-    private String descricao;
-
-    public Comentario(int id, String nome, String descricao) { // Construtor atualizado
-        this.id = id;
-        this.nome = nome;
-        this.descricao = descricao;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getDescricao() {
-        return descricao;
-    }
-
-    public void setDescricao(String descricao) {
-        this.descricao = descricao;
-    }
-}
