+++ src/Frontend/app/src/main/java/br/fecap/pi/ludis/FeedSimulator.java
@@ -1,0 +1,39 @@
+package br.fecap.pi.ludis;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.widget.ImageView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.graphics.Insets;
+import androidx.core.view.ViewCompat;
+import androidx.core.view.WindowInsetsCompat;
+
+public class FeedSimulator extends AppCompatActivity {
+
+    private ImageView btnVoltar;
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_telafeed);
+
+        Intent inte = getIntent();
+        String nome = inte.getStringExtra("nome");
+
+
+        btnVoltar = findViewById(R.id.btnVoltar);
+        btnVoltar.setOnClickListener(view -> {
+            Intent voltarIntent = new Intent(FeedSimulator.this, Principal.class);
+            startActivity(voltarIntent);
+            finish();
+        });
+
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
+            Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
+            return insets;
+        });
+    }
+}
